changes required by transaction

+++ CleanArchitecture.Core/CleanArchitecture.Core/ApiModels/BizResponseClass.cs
@@ -6,13 +6,13 @@ namespace CleanArchitecture.Core.ApiModels
 {
     public class BizResponseClass
     {
-        public byte ReturnCode { get; set; }
+        public short ReturnCode { get; set; }
 
         public string ReturnMsg { get; set; }
 
-        public int ErrorCode { get; set; }
+        public short ErrorCode { get; set; }
 
-        public byte StatusCode { get; set; }
+        public short StatusCode { get; set; }
 
         public string StatusMessage { get; set; }
